tests/wpt/tests/css/css-scroll-snap-2/parsing/scroll-start-target-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap 2 Test: scroll-none-target sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-2/#scroll-start-target">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>

<body>
  <script>
    test_shorthand_value("scroll-start-target", "none", {
      "scroll-start-target-block": "none",
      "scroll-start-target-inline": "none",
    });
    test_shorthand_value("scroll-start-target", "auto", {
      "scroll-start-target-block": "auto",
      "scroll-start-target-inline": "none",
    });
    test_shorthand_value("scroll-start-target", "none auto", {
      "scroll-start-target-block": "none",
      "scroll-start-target-inline": "auto",
    });
    test_shorthand_value("scroll-start-target", "auto none", {
      "scroll-start-target-block": "auto",
      "scroll-start-target-inline": "none",
    });
  </script>
</body>
